New Table value adjustments
Actually ensure that if a gun, bomb, or pod is added, and if the addition is missing data, it has a fallback table to reference.

Special note, in Lua 5.1, ipairs and the table library does not yet respect metamethods.
For example, table.insert does not invoke __newIndex.

Additional HUD Elements